Smart wallet selectors refactoring
