+++ src/selectors/selectors.js
@@ -21,11 +21,8 @@
 import get from 'lodash.get';
 import { createSelector } from 'reselect';
 import { getAccountAddress } from 'utils/accounts';
-import { accountHasGasTokenSupport, getSmartWalletStatus } from 'utils/smartWallet';
-import { SMART_WALLET_UPGRADE_STATUSES } from 'constants/smartWalletConstants';
 
 import type { RootReducerState } from 'reducers/rootReducer';
-import type { SmartWalletStatus } from 'models/SmartWalletStatus';
 
 //
 // Global selectors here
@@ -75,15 +72,3 @@ export const activeBlockchainSelector = ({ appSettings }: RootReducerState) =>
   get(appSettings, 'data.blockchainNetwork', 'Ethereum');
 
 export const featureFlagsSelector = ({ featureFlags }: RootReducerState) => featureFlags.data;
-
-export const isSmartWalletActivatedSelector = ({
-  accounts: { data: accounts },
-  smartWallet,
-}: RootReducerState) => {
-  const smartWalletStatus: SmartWalletStatus = getSmartWalletStatus(accounts, smartWallet);
-  return (smartWalletStatus.status === SMART_WALLET_UPGRADE_STATUSES.DEPLOYMENT_COMPLETE);
-};
-
-export const isGasTokenSupportedSelector = ({
-  smartWallet: { connectedAccount },
-}: RootReducerState) => accountHasGasTokenSupport(connectedAccount);
